+++ app/src/main/java/com/example/expenditure/Login/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.expenditure;
+package com.example.expenditure.Login;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -10,6 +10,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.expenditure.R;
+import com.example.expenditure.ScrollingActivity;
 import com.google.android.gms.auth.api.signin.GoogleSignIn;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.android.gms.auth.api.signin.GoogleSignInClient;
