+++ app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginActivity"
+    tools:context=".Login.LoginActivity"
     android:id="@+id/constraint_layout_activity_login">
 
     <com.google.android.gms.common.SignInButton
